% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_matrix}
\alias{ensure_matrix}
\title{Convertit les données en matrice numérique et vérifie les dimensions}
\usage{
ensure_matrix(X, y)
}
\description{
Convertit les données en matrice numérique et vérifie les dimensions
}
