% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{new_node}
\alias{new_node}
\title{Construction d'un nœud AID (structure S3)}
\usage{
new_node(
  depth,
  n,
  sse,
  mean,
  parent_value = NA_real_,
  feature = NA_integer_,
  threshold = NA_real_,
  gain = 0,
  f_stat = NA_real_,
  left = NULL,
  right = NULL
)
}
\description{
Construction d'un nœud AID (structure S3)
}
