% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_count.R
\name{raid_barcode_count}
\alias{raid_barcode_count}
\title{Count the number of unique UMI per Barcode_1 and Barcode_2 columns.
The data table that should be used is the iSeq_UMI_count table, which
contains all the unique (collapsed) UMI,
barcode_1 and barcode_2 information. The column row_occurence_count can be
ignored.}
\usage{
raid_barcode_count(split_data_file_path, UMI_count_tbl = UMI_count,
  output_filename = "output/data/barcode_count_")
}
\description{
Count the number of unique UMI per Barcode_1 and Barcode_2 columns.
The data table that should be used is the iSeq_UMI_count table, which
contains all the unique (collapsed) UMI,
barcode_1 and barcode_2 information. The column row_occurence_count can be
ignored.
}

