% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI_analysis.R
\name{raid_import_split_data}
\alias{raid_import_split_data}
\title{# Function for UMI analysis from iSeq split data table.
 sequence. The first question is: how many PCR duplicates can be found?
 A PCR duplicate will result a duplicate UMI (and barcodes) sequences.
 Practically this means if a row is a similar to another one, this is a PCR
 duplication.}
\usage{
raid_import_split_data(split_data_file_path)
}
\description{
# Count the duplicate UMI in the whole file & filter duplicates.
Three functions are defined:
1. import the split reads into the R environment
2. count the occurence of a specific row (which contains 1 read)
3. count the frequency of the occurences (direct usable for UMI frequency
   plot)
}
\details{
Function 2 and 3 each save a TSV file in the output/data/ folder .
}

