% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reads_from_FASTQ.R
\name{raid_split_reads}
\alias{raid_split_reads}
\title{Split reads from FASTQ files}
\usage{
raid_split_reads(fastq_file)
}
\arguments{
\item{fastq_file}{fastq formatted files in a directory path. Default settings or more information from ShortRead::readFastq(fastq_file)}
}
\value{
a tsv file with three columns: UMI , Barcode_1 , sample_ID
}
\description{
First, the iseq_split_reads function uses shortread package function "readFastq" to load all reads from
a .fastq or .fastq.gz file. 
Then the reads containing an anchor sequence are extracted 
and these reads are split to retreive - UMI - Barcode_1 sequences (and stored in a table with 2 columns)
Finally a column with the sample_ID is added.
}
\examples{
RAID_split_reads("experiment/data/sample-ID.fastq.gz")
}
\seealso{
\code{\link{ShortRead::readFastq}}
}

